+++ src/components/common/query/sidebar/panels/ShareSnapshot.jsx
